requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.18.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# Vector stores and embeddings
chromadb>=0.4.22
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Document processing
pypdf>=3.17.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
unstructured>=0.12.0
tiktoken>=0.5.2

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# API and web interface
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.30.0
python-multipart>=0.0.6

# Evaluation
ragas>=0.1.0
datasets>=2.16.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
tqdm>=4.66.0
